the collision seems to work up to 100000 blocks

+++ include/gameLayer/gameplay/physics.h
@@ -17,7 +17,7 @@ struct RigidBody
 	void checkCollisionBrute(glm::dvec3 &pos, glm::dvec3 lastPos, 
 		decltype(chunkGetterSignature) *chunkGetter);
 
-	glm::dvec3 performCollision(glm::dvec3 pos, glm::dvec3 lastPos, glm::vec3 size, glm::vec3 delta,
+	glm::dvec3 performCollision(glm::dvec3 pos, glm::dvec3 lastPos, glm::vec3 size, glm::dvec3 delta,
 		decltype(chunkGetterSignature) *chunkGetter);
 
 	void updateMove();
@@ -27,6 +27,8 @@ struct RigidBody
 //this is the local player
 struct Player
 {
+
+	//todo will probably remove the idea of a rigid body
 	glm::vec3 lookDirection = {0,0,-1};
 	RigidBody body = {};
 	std::uint64_t entityId = 0;
